refactor: extract Block model into its own Django app
Move Block model, is_blocked helper, and block/unblock views out of the friends app into a dedicated block app.
Update chat consumers, settings, urls, and frontend API calls accordingly.
- Frontend calls /api/block/ (POST) and /api/block/unblock (DELETE)
- friends/migrations/0002_block.py removed; block/migrations/0001_initial.py added
- Document Block REST API in README
