+++ backend/django_server/urls.py
@@ -4,6 +4,7 @@ urlpatterns = [
     path('api/auth/', include('users.urls')),
     path('api/tournament/', include('tournament.urls')),
     path('api/friends/', include('friends.urls')),
+    path('api/block/', include('block.urls')),
     path('api/chess/', include('chessgame.urls')),
     path('api/', include('game.urls')),
