+++ backend/friends/urls.py
@@ -7,6 +7,4 @@ urlpatterns = [
 	path('delete', views.delete_request, name='delete_request'),
 	path('list', views.get_friends, name='get_friends_list'),
 	path('remove', views.remove_friend, name='remove_friend'),
-	path('block', views.block_user, name='block_user'),
-	path('unblock', views.unblock_user, name='unblock_user'),
 ]
